move python code to build info go
