+++ xray/commands/audit/python/python_test.go
@@ -1,6 +1,7 @@
 package python
 
 import (
+	"github.com/jfrog/jfrog-cli-core/v2/artifactory/utils"
 	"github.com/jfrog/jfrog-cli-core/v2/xray/commands/audit"
 	"github.com/stretchr/testify/assert"
 	"path/filepath"
@@ -12,8 +13,8 @@ func TestBuildPipDependencyListSetuppy(t *testing.T) {
 	_, cleanUp := audit.CreateTestWorkspace(t, filepath.Join("pip-project", "setuppyproject"))
 	defer cleanUp()
 	// Run getModulesDependencyTrees
-	auditCmd := NewEmptyAuditPipCommand()
-	dependencyTree, err := auditCmd.buildPipDependencyTree()
+	auditCmd := NewEmptyAuditPythonCommand(utils.Pip)
+	dependencyTree, err := auditCmd.buildDependencyTree()
 	if assert.NoError(t, err) && assert.NotNil(t, dependencyTree) {
 		assert.NotEmpty(t, dependencyTree.Nodes)
 		// Test root module
@@ -30,8 +31,8 @@ func TestBuildPipDependencyListRequirements(t *testing.T) {
 	_, cleanUp := audit.CreateTestWorkspace(t, filepath.Join("pip-project", "requirementsproject"))
 	defer cleanUp()
 	// Run getModulesDependencyTrees
-	auditCmd := NewEmptyAuditPipCommand()
-	rootNodes, err := auditCmd.buildPipDependencyTree()
+	auditCmd := NewEmptyAuditPythonCommand(utils.Pip)
+	rootNodes, err := auditCmd.buildDependencyTree()
 	if assert.NoError(t, err) && assert.NotEmpty(t, rootNodes) {
 		if rootNodes != nil {
 			// Test root module
@@ -39,5 +40,25 @@ func TestBuildPipDependencyListRequirements(t *testing.T) {
 			// Test child module
 			audit.GetAndAssertNode(t, rootNode.Nodes, "ptyprocess:0.7.0")
 		}
+	}
+}
+
+func TestBuildPipenvDependencyList(t *testing.T) {
+	// Create and change directory to test workspace
+	_, cleanUp := audit.CreateTestWorkspace(t, "pipenv-project")
+	defer cleanUp()
+	// Run getModulesDependencyTrees
+	auditCmd := NewEmptyAuditPythonCommand(utils.Pipenv)
+	dependencyTree, err := auditCmd.buildDependencyTree()
+	if assert.NoError(t, err) && assert.NotNil(t, dependencyTree) {
+		assert.NotEmpty(t, dependencyTree.Nodes)
+		// Test child module
+		audit.GetAndAssertNode(t, dependencyTree.Nodes, "toml:0.10.2")
+
+		// Test child module
+		childNode := audit.GetAndAssertNode(t, dependencyTree.Nodes, "pexpect:4.8.0")
+		// Test sub child module
+		audit.GetAndAssertNode(t, childNode.Nodes, "ptyprocess:0.7.0")
+
 	}
 }
